Added Tools application for maintenance
