+++ dashboard/templates/dashboard/nav_csd_atelier.html
@@ -76,12 +76,12 @@
                 <div class="bg-white py-2 collapse-inner rounded">
                     <h6 class="collapse-header">Informations :</h6>
                     <a class="collapse-item"
-                       href="{% url 'dashboard:soft-list' %}">{% trans "Software list" %}</a>
+                       href="{% url 'tools:soft-list' %}">{% trans "Software list" %}</a>
                     <div class="collapse-divider"></div>
                     {% if user|has_group:'cellule' %}
                         <h6 class="collapse-header">Edition :</h6>
                         <a class="collapse-item"
-                           href="{% url 'dashboard:soft-add' %}">{% trans "Software Add" %}</a>
+                           href="{% url 'tools:soft-add' %}">{% trans "Software Add" %}</a>
                     {% endif %}
                 </div>
             </div>
