+++ dashboard/tests/tests_dashboard_views.py
@@ -3,16 +3,12 @@ from django.urls import reverse
 from django.utils import translation
 from django.contrib.auth.models import User, Group
 
-from dashboard.models import CsdSoftware
 from squalaetp.models import Xelon
 
 
 class DashboardTestCase(TestCase):
 
     def setUp(self):
-        self.form_data = {
-            'jig': 'test', 'new_version': '1', 'link_download': 'test', 'status': 'En test',
-        }
         self.vin = 'VF3ABCDEF12345678'
         user = User.objects.create_user(username='toto', email='toto@bibi.com', password='totopassword')
         user.groups.add(Group.objects.create(name="cellule"))
@@ -47,28 +43,6 @@ class DashboardTestCase(TestCase):
         response = self.client.get(reverse('dashboard:register'))
         self.assertEqual(response.status_code, 302)
         self.assertRedirects(response, '/accounts/login/?next=/dashboard/register/')
-
-    def test_soft_list_page(self):
-        response = self.client.get(reverse('dashboard:soft-list'))
-        self.assertEqual(response.status_code, 200)
-
-    def test_soft_add_page_is_disconnected(self):
-        response = self.client.get(reverse('dashboard:soft-add'))
-        self.assertEqual(response.status_code, 302)
-        self.assertRedirects(response, '/accounts/login/?next=/dashboard/soft/add/')
-
-    def test_soft_add_page_is_connected(self):
-        self.client.login(username='toto', password='totopassword')
-        response = self.client.get(reverse('dashboard:soft-add'))
-        self.assertEqual(response.status_code, 200)
-
-    def test_soft_add_page_is_valid(self):
-        self.client.login(username='toto', password='totopassword')
-        old_soft = CsdSoftware.objects.count()
-        response = self.client.post(reverse('dashboard:soft-add'), self.form_data)
-        new_soft = CsdSoftware.objects.count()
-        self.assertEqual(new_soft, old_soft + 1)
-        self.assertEqual(response.status_code, 200)
 
     def test_search_is_not_found(self):
         response = self.client.get(reverse('dashboard:search'), {'query': 'null', 'select': 'xelon'},
