+++ dashboard/views.py
@@ -9,8 +9,8 @@ import re
 
 from utils.product_Analysis import ProductAnalysis
 from utils.decorators import group_required
-from .models import Post, CsdSoftware, User, UserProfile
-from .forms import SoftwareForm, ParaErrorList, UserProfileForm, SignUpForm
+from .models import Post, UserProfile
+from .forms import UserProfileForm, SignUpForm
 from squalaetp.models import Xelon
 
 
@@ -106,70 +106,6 @@ def register(request):
         form = SignUpForm()
     context['form'] = form
     return render(request, 'registration/register.html', context)
-
-
-def soft_list(request):
-    """
-    View of Software list page
-    """
-    softs = CsdSoftware.objects.all()
-    context = {
-        'title': 'Software',
-        'table_title': _('Software list'),
-        'softs': softs,
-    }
-    return render(request, 'dashboard/soft_table.html', context)
-
-
-@login_required
-@group_required('cellule')
-def soft_add(request):
-    """
-    View for adding a software in the list
-    """
-    context = {
-        'title': 'Software',
-        'card_title': _('Software integration'),
-    }
-    if request.method == 'POST':
-        user = User.objects.get(pk=request.user.id)
-        form = SoftwareForm(request.POST, error_class=ParaErrorList)
-        if form.is_valid():
-            jig = form.cleaned_data['jig']
-            ref = CsdSoftware.objects.filter(jig=jig)
-            if not ref.exists():
-                CsdSoftware.objects.create(**form.cleaned_data, created_by=user)
-                context = {'title': _('Added successfully!')}
-                return render(request, 'dashboard/done.html', context)
-        context['errors'] = form.errors.items()
-    else:
-        form = SoftwareForm()
-    context['form'] = form
-    return render(request, 'dashboard/soft_add.html', context)
-
-
-@login_required
-@group_required('cellule')
-def soft_edit(request, soft_id):
-    """
-    View for changing software data
-    :param soft_id:
-        Software id to edit
-    """
-    soft = get_object_or_404(CsdSoftware, pk=soft_id)
-    form = SoftwareForm(request.POST or None, instance=soft)
-    if form.is_valid():
-        form.save()
-        context = {'title': _('Modification done successfully!')}
-        return render(request, 'dashboard/done.html', context)
-    context = {
-        'title': 'Software',
-        'card_title': _('Modification data Software for JIG: {jig}'.format(jig=soft.jig)),
-        'url': 'dashboard:soft-edit',
-        'soft': soft,
-        'form': form,
-    }
-    return render(request, 'dashboard/soft_edit.html', context)
 
 
 @login_required
