+++ sbadmin/settings/__init__.py
@@ -35,6 +35,7 @@ INSTALLED_APPS = [
     'squalaetp.apps.SqualaetpConfig',
     'reman.apps.RemanConfig',
     'demo.apps.DemoConfig',
+    'tools.apps.ToolsConfig',
     'api.apps.ApiConfig',
     'django.contrib.admin',
     'django.contrib.auth',
