+++ tools/templates/tools/soft_add.html
@@ -5,9 +5,9 @@
 {% block form_content %}
 
     <div class="p-2">
-        <form method="post" action="{% url 'dashboard:soft-add' %}">
+        <form method="post" action="{% url 'tools:soft-add' %}">
 
-            {% include 'dashboard/soft_form.html' %}
+            {% include 'tools/soft_form.html' %}
 
         </form>
